Switch From Tomcat to Undertow
This has to be done since keycloak
uses a security class that is no longer
present when using tomcat with the targeted
java version. However, using Undertow or
Jetty this class is still present.
A fix for this behavior is going to land
in keycloak 13. For further information
about this issue follow the link:
https://github.com/keycloak/keycloak/pull/7533