Refactor Query Result Collect Process
Removes any need for keeping track of different broker clients
and their associated external queries. As soon as an update on
a query is present, the associated broker gets forwarded along
other necessary information identifying the external query
within said broker. This enables the receiving end to communicate
with the broker on an ad-hoc basis.

This also removes the need for keeping track of external identifiers
throughout different places within the database. Instead, a central
table is used for this very purpose.