Ensure Result Only Affected by Successful Queries
Ensures that a requested result summary is only affected
by successful queries.
Queries terminated by an error will not be part of the
summary.

The former behavior would only ever be useful if the frontend
was displaying a query to a specific site as failed. Since
there is currently no plan to do so, the assertions above have
been enforced.