Reimplement Site Name Obfuscation
Reimplements site name obfuscation based on a tokenization technique.
This technique replaces the actual site name with a placeholder that
is called a token.
A token for a site name to a corresponding result gets calculated
using a hash function. Eventually the output of said hash function is
truncated to a length that is sufficient for this project.

As requested in the initial issue #49 this reimplementation is stable.
I.e. multiple calls to tokenize a site name of a given result will
yield the same token without this operation being reversible.