Fix Race Condition Using Multiple Brokers
Fixes a race condition that can occur when using
multiple brokers.
For the race condition to occur, one of the brokers
must take a long time to publish a query while
another one already managed to produce a result. This
would lead to a missing database entry since the
dispatch function is transactional over all brokers
and does not commit any transaction until all brokers
have published their query.

In order to prevent this situation from happening, the
database is not used for any sort of component
communication anymore. Instead necessary broker and
backend query ID translations are moved to the broker
implementations.