Support Latest Flare Version
Adds support for the latest Flare version in
development v1.0. These changes support Flare
in version of at least v1.0-RC6.

Mainly adjusts communication headers since in
the past Flare enforced the presence of headers
in an incorrect fashion (headers were used for
things that they were not supposed to).