Move and Rename to Clarify Association with API
This is simply done in order to group what needs
to go together. Additionally, renames several
classes so that they have a StructuredQuery prefix.
Without this adjustment we are running into a
mental naming clash since there is also the concept
of an internal query. Having the prefix clarifies
the context.