Respect Dispatches After First Successful One
Adjusts dispatch behavior so that after a successful first
dispatch, following dispatches are respected as well.
The former implementation had some form of short-circuit in
this regard. This led to brokers not being triggered to
publish a query. This only happened if multiple brokers where
configured and the first one was also successful.